Extract image-specific rebuild code to new module
